requirements.txt
flask==2.3.3
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
werkzeug==2.3.7
python-decouple==3.8
requests==2.31.0
langchain
langchain-groq
langchain-core
langchain-chroma
langchain-huggingface
chromadb==0.4.17
stripe==7.4.0
gunicorn==21.2.0
flask-migrate==4.0.4
